Hacky first implementation of TPM support for hidden services
Use a TPM-based key when bringing up hidden services. This should be entirely
invisible to users as long as they're running Trousers on a version 1.2 TPM.
Note that it currently re-uses the existing hidden services private key
filename but uses an entirely different format, and also that keys can't be
imported and so your hidden service names will be different.